requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.1
django-ckeditor==6.5.1
django-debug-toolbar==4.0.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-multiupload==0.6.1
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-recaptcha3==0.4.0
django-reversion==5.0.4
idna==3.4
phonenumbers==8.13.8
phonenumberslite==8.13.8
Pillow==9.5.0
pytz==2023.3
requests==2.30.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
urllib3==2.0.2
